https://github.com/jackdewinter/pymarkdown/issues/168 - Moved paragraph code to own module.
