+++ pymarkdown/leaf_block_processor.py
@@ -10,7 +10,6 @@ from pymarkdown.inline_helper import InlineHelper
 from pymarkdown.inline_markdown_token import TextMarkdownToken
 from pymarkdown.leaf_markdown_token import (
     AtxHeadingMarkdownToken,
-    BlankLineMarkdownToken,
     FencedCodeBlockMarkdownToken,
     IndentedCodeBlockMarkdownToken,
     ParagraphMarkdownToken,
@@ -23,7 +22,6 @@ from pymarkdown.parser_logger import ParserLogger
 from pymarkdown.parser_state import ParserState
 from pymarkdown.position_marker import PositionMarker
 from pymarkdown.stack_token import (
-    BlockQuoteStackToken,
     FencedCodeBlockStackToken,
     IndentedCodeBlockStackToken,
     ListStackToken,
@@ -290,20 +288,6 @@ class LeafBlockProcessor:
     # pylint: enable=too-many-arguments
 
     @staticmethod
-    def __adjust_for_list_start(
-        parser_state: ParserState,
-        last_list_start_index: int,
-        last_block_quote_index: int,
-    ) -> int:
-        # TODO remove?
-        POGGER.debug("last_list_start_index>>$>>", last_list_start_index)
-        POGGER.debug(
-            "parser_state.original_line_to_parse>>$>>",
-            parser_state.original_line_to_parse,
-        )
-        return last_block_quote_index
-
-    @staticmethod
     def __recalculate_whitespace(
         whitespace_to_parse: Optional[str], offset_index: int
     ) -> Tuple[int, str, str]:
@@ -324,7 +308,6 @@ class LeafBlockProcessor:
 
         return actual_whitespace_index, adj_ws, left_ws
 
-    # pylint: disable=too-many-arguments
     @staticmethod
     def parse_indented_code_block(
         parser_state: ParserState,
@@ -332,7 +315,6 @@ class LeafBlockProcessor:
         extracted_whitespace: Optional[str],
         removed_chars_at_start: Optional[int],
         last_block_quote_index: int,
-        last_list_start_index: int,
     ) -> List[MarkdownToken]:
 
         """
@@ -355,7 +337,6 @@ class LeafBlockProcessor:
                     extracted_whitespace,
                 ) = LeafBlockProcessor.__create_indented_block(
                     parser_state,
-                    last_list_start_index,
                     last_block_quote_index,
                     position_marker,
                     extracted_whitespace,
@@ -372,13 +353,9 @@ class LeafBlockProcessor:
             )
         return new_tokens
 
-    # pylint: enable=too-many-arguments
-
-    # pylint: disable=too-many-arguments
     @staticmethod
     def __create_indented_block(
         parser_state: ParserState,
-        last_list_start_index: int,
         last_block_quote_index: int,
         position_marker: PositionMarker,
         extracted_whitespace: Optional[str],
@@ -399,7 +376,6 @@ class LeafBlockProcessor:
             extracted_whitespace,
         ) = LeafBlockProcessor.__prepare_for_indented_block(
             parser_state,
-            last_list_start_index,
             last_block_quote_index,
             extracted_whitespace,
         )
@@ -414,22 +390,12 @@ class LeafBlockProcessor:
 
         return last_block_quote_index, extracted_whitespace
 
-    # pylint: enable=too-many-arguments
-
     @staticmethod
     def __prepare_for_indented_block(
         parser_state: ParserState,
-        last_list_start_index: int,
         last_block_quote_index: int,
         extracted_whitespace: Optional[str],
     ) -> Tuple[int, int, str, str]:
-        POGGER.debug(">>__adjust_for_list_start")
-        (last_block_quote_index) = LeafBlockProcessor.__adjust_for_list_start(
-            parser_state,
-            last_list_start_index,
-            last_block_quote_index,
-        )
-
         POGGER.debug(">>>>$", parser_state.token_stack[-1])
         if parser_state.token_stack[-1].is_list:
             list_token = cast(ListStackToken, parser_state.token_stack[-1])
@@ -865,194 +831,6 @@ class LeafBlockProcessor:
     # pylint: enable=too-many-arguments
 
     @staticmethod
-    def parse_paragraph(
-        parser_state: ParserState,
-        position_marker: PositionMarker,
-        extracted_whitespace: Optional[str],
-        block_quote_data: BlockQuoteData,
-        text_removed_by_container: Optional[str],
-    ) -> List[MarkdownToken]:
-        """
-        Handle the parsing of a paragraph.
-        """
-        POGGER.debug(">>text_removed_by_container>>:$:<<", text_removed_by_container)
-        assert extracted_whitespace is not None
-        if parser_state.no_para_start_if_empty and position_marker.index_number >= len(
-            position_marker.text_to_parse
-        ):
-            POGGER.debug("Escaping paragraph due to empty w/ blank")
-            POGGER.debug(
-                "position_marker.text_to_parse=:$:", position_marker.text_to_parse
-            )
-            POGGER.debug(
-                "position_marker.index_number=:$:", position_marker.index_number
-            )
-            POGGER.debug(
-                "position_marker.index_indent=:$:", position_marker.index_indent
-            )
-            return [
-                BlankLineMarkdownToken(
-                    extracted_whitespace, position_marker, len(extracted_whitespace)
-                )
-            ]
-
-        POGGER.debug(
-            "parse_paragraph>block_quote_data.stack_count>$>block_quote_data.current_count>$<",
-            block_quote_data.stack_count,
-            block_quote_data.current_count,
-        )
-
-        container_index, adjusted_whitespace_length = (
-            parser_state.find_last_container_on_stack(),
-            0,
-        )
-        if container_index > 0:
-            adjusted_whitespace_length = (
-                LeafBlockProcessor.__adjust_paragraph_for_containers(
-                    parser_state,
-                    container_index,
-                    extracted_whitespace,
-                    adjusted_whitespace_length,
-                )
-            )
-
-        new_tokens, extracted_whitespace = LeafBlockProcessor.__handle_paragraph_prep(
-            parser_state,
-            block_quote_data,
-            position_marker,
-            extracted_whitespace,
-        )
-
-        assert extracted_whitespace is not None
-        new_tokens.append(
-            TextMarkdownToken(
-                position_marker.text_to_parse[position_marker.index_number :],
-                extracted_whitespace,
-                position_marker=position_marker,
-            )
-        )
-        return new_tokens
-
-    @staticmethod
-    def __handle_paragraph_prep(
-        parser_state: ParserState,
-        block_quote_data: BlockQuoteData,
-        position_marker: PositionMarker,
-        extracted_whitespace: Optional[str],
-    ) -> Tuple[List[MarkdownToken], Optional[str]]:
-
-        # In cases where the list ended on the same line as we are processing, the
-        # container tokens will not yet be added to the token_document.  As such,
-        # make sure to construct a "proper" list that takes those into account
-        # before checking to see if this is an issue.
-        adjusted_document = parser_state.token_document[:]
-        assert parser_state.same_line_container_tokens is not None
-        adjusted_document.extend(parser_state.same_line_container_tokens)
-
-        if (
-            len(adjusted_document) >= 2
-            and adjusted_document[-1].is_blank_line
-            and adjusted_document[-2].is_any_list_token
-        ):
-
-            did_find, last_list_index = LeafBlockProcessor.check_for_list_in_process(
-                parser_state
-            )
-            assert did_find
-            new_tokens, _ = parser_state.close_open_blocks_fn(
-                parser_state, until_this_index=last_list_index
-            )
-        elif block_quote_data.stack_count != 0 and block_quote_data.current_count == 0:
-            new_tokens, _ = parser_state.close_open_blocks_fn(
-                parser_state,
-                only_these_blocks=[BlockQuoteStackToken],
-                include_block_quotes=True,
-            )
-        else:
-            new_tokens = []
-
-        if not parser_state.token_stack[-1].is_paragraph:
-            assert extracted_whitespace is not None
-            new_paragraph_token = ParagraphMarkdownToken(
-                extracted_whitespace, position_marker
-            )
-            parser_state.token_stack.append(ParagraphStackToken(new_paragraph_token))
-            new_tokens.append(new_paragraph_token)
-            extracted_whitespace = ""
-        return new_tokens, extracted_whitespace
-
-    @staticmethod
-    def __adjust_paragraph_for_containers(
-        parser_state: ParserState,
-        container_index: int,
-        extracted_whitespace: Optional[str],
-        adjusted_whitespace_length: int,
-    ) -> int:
-        if not parser_state.token_stack[container_index].is_block_quote:
-            top_list_token = cast(
-                ListStackToken, parser_state.token_stack[container_index]
-            )
-            POGGER.debug(">>list-owners>>$", top_list_token)
-            adjusted_whitespace_length = LeafBlockProcessor.__adjust_paragraph_for_list(
-                top_list_token, extracted_whitespace
-            )
-        POGGER.debug(">>adjusted_whitespace_length>>$", adjusted_whitespace_length)
-        return adjusted_whitespace_length
-
-    @staticmethod
-    def __adjust_paragraph_for_list(
-        top_list_token: ListStackToken, extracted_whitespace: Optional[str]
-    ) -> int:
-        assert extracted_whitespace is not None
-        ex_ws_length = len(extracted_whitespace)
-        POGGER.debug(">>owners-indent>>$", top_list_token.indent_level)
-        POGGER.debug(">>ws_before_marker>>$", top_list_token.ws_before_marker)
-        POGGER.debug(">>ws_after_marker>>$", top_list_token.ws_after_marker)
-        POGGER.debug(">>last_new_list_token>>$", top_list_token.last_new_list_token)
-        POGGER.debug(">>extracted_whitespace>>$", ex_ws_length)
-
-        dominant_indent, original_list_indent, indent_delta = (
-            (
-                top_list_token.last_new_list_token.indent_level
-                if top_list_token.last_new_list_token
-                else top_list_token.indent_level
-            ),
-            top_list_token.indent_level - 2,
-            (
-                top_list_token.ws_after_marker - 1
-                if top_list_token.ws_after_marker > 1
-                else 0
-            ),
-        )
-        POGGER.debug(">>dominant_indent>>$>>", dominant_indent)
-        original_text_indent = (
-            ex_ws_length
-            + top_list_token.indent_level
-            - top_list_token.ws_before_marker
-            - indent_delta
-        )
-        POGGER.debug(">>original_list_indent>>$>>", original_list_indent)
-        POGGER.debug(">>original_text_indent>$>>", original_text_indent)
-        return (
-            dominant_indent - original_text_indent
-            if dominant_indent > original_text_indent >= 4
-            else 0
-        )
-
-    @staticmethod
-    def check_for_list_in_process(parser_state: ParserState) -> Tuple[bool, int]:
-        """
-        From the end of the stack, check to see if there is already a list in progress.
-        """
-
-        stack_index = len(parser_state.token_stack) - 1
-
-        while stack_index >= 0 and not parser_state.token_stack[stack_index].is_list:
-            stack_index -= 1
-
-        return stack_index >= 0, stack_index
-
-    @staticmethod
     def correct_for_leaf_block_start_in_list(
         parser_state: ParserState,
         removed_chars_at_start: int,
@@ -1174,10 +952,6 @@ class LeafBlockProcessor:
         Determine whether we have a valid leaf block start.
         """
 
-        # TODO Can be Removed?
-        POGGER.debug(
-            "is_paragraph_ending_leaf_block_start, ex=$", exclude_thematic_break
-        )
         is_leaf_block_start = not exclude_thematic_break
         assert not exclude_thematic_break
